+++ src/main/java/com/ghostchu/btn/sparkle/module/tracker/TrackerService.java
@@ -243,8 +243,8 @@ public class TrackerService {
             PeerEvent peerEvent,
             String userAgent,
             boolean supportCrypto, String key, long corrupt, long redundant,
-            String trackerId, int cryptoPort, boolean hide, int azudp, int azhttp,
-            boolean azq, String azver, int azup, String azas, String aznp, long numWant) implements Serializable {
+            String trackerId, long cryptoPort, boolean hide, long azudp, long azhttp,
+            boolean azq, String azver, long azup, String azas, String aznp, long numWant) implements Serializable {
 
     }
